compiling with filter
