+++ ownership-chain/runtime/laos/src/configs/mod.rs
@@ -5,6 +5,7 @@ mod block_rewards_source;
 mod multisig;
 pub mod parachain_staking;
 mod proxy;
+mod system;
 mod utility;
 
 // Define storage fees as constants for clarity and reuse
